updates and fixes os-setup.sh

+++ os-setup.sh
@@ -15,7 +15,7 @@ gsettings set org.freedesktop.ibus.panel.emoji unicode-hotkey "[]"
 
 function clear_fs_with() {
     local mod=$1
-    gsettings list-recursively | grep "'<$mod>F[0-9][0-9]" | while read -r line; do
+    gsettings list-recursively | grep "'<$mod>F[0-9][0-9]*" | while read -r line; do
         new=$(echo "$line" | sed -e "s/.*\[/[/g" -e "s/, '<$mod>F[0-9][0-9]*'//" -e "s/'<$mod>F[0-9][0-9]*',*//")
         group=$(echo "$line" | awk '{ print $1 }')
         key=$(echo "$line" | awk '{ print $2 }')
@@ -28,9 +28,8 @@ clear_fs_with "Alt"
 clear_fs_with "Super"
 
 for i in $(seq 4); do
-    gsettings set org.gnome.desktop.wm.keybindings switch-to-workspace-1 "['<Super>F$i']"
+    gsettings set org.gnome.desktop.wm.keybindings "switch-to-workspace-$i" "['<Super>F$i']"
 done
-
 
 sudo apt-get update
 sudo apt-get dist-upgrade
@@ -38,9 +37,13 @@ sudo apt-get autoremove
 
 sudo apt-get install -y vim openssh-server git default-jdk maven postgresql docker docker-compose curl
 
-curl -L curl -L https://download.jetbrains.com/idea/ideaIU-2019.3.4.tar.gz > /tmp/idea.tar.gz
-cd /opt || exit 1
-rm -rf idea*
-sudo tar xfz /tmp/idea.tar.gz
-sudo chown "$(id -g):$(id -u)" idea*
-sudo update-alternatives --install /usr/local/bin/idea idea /opt/idea*/bin/idea.sh 1
+if hash idea 2> /dev/null; then
+    echo "Idea already installed"
+else
+    curl -L curl -L https://download.jetbrains.com/idea/ideaIU-2019.3.4.tar.gz > /tmp/idea.tar.gz
+    cd /opt || exit 1
+    rm -rf idea*
+    sudo tar xfz /tmp/idea.tar.gz
+    sudo chown "$(id -g):$(id -u)" idea*
+    sudo update-alternatives --install /usr/local/bin/idea idea /opt/idea*/bin/idea.sh 1
+fi
